Fixed ministerial score to be float64, not int64

+++ internal/requests/bibliometrics_requests.go
@@ -17,7 +17,7 @@ type CreateBibliometric struct {
 	HIndexWos        int       `query:"h_index_wos" json:"h_index_wos" doc:"HIndex Wos" format:"int" example:"1"`
 	HIndexScopus     int       `query:"h_index_scopus" json:"h_index_scopus" doc:"HIndex Scopus" format:"int" example:"2"`
 	PublicationCount int       `query:"publication_count" json:"publication_count" doc:"Publication count" format:"int" example:"123"`
-	MinisterialScore float64   `query:"ministerial_score" json:"ministerial_score" doc:"Ministerial score" format:"float" example:"65.7"`
+	MinisterialScore float64   `query:"ministerial_score" json:"ministerial_score" doc:"Ministerial score" format:"float64" example:"65.7"`
 	ScientistID      uuid.UUID `query:"scientist_id" json:"scientist_id" doc:"Scientist ID" format:"uuid" example:"8c4bfb01-3c0a-416c-a07c-a24ee52a8b2a"`
 }
 type UpdateBibliometric struct {
@@ -25,7 +25,7 @@ type UpdateBibliometric struct {
 	HIndexWos        int       `query:"h_index_wos" json:"h_index_wos" doc:"HIndex Wos" format:"int" example:"1"`
 	HIndexScopus     int       `query:"h_index_scopus" json:"h_index_scopus" doc:"HIndex Scopus" format:"int" example:"2"`
 	PublicationCount int       `query:"publication_count" json:"publication_count" doc:"Publication count" format:"int" example:"123"`
-	MinisterialScore float64   `query:"ministerial_score" json:"ministerial_score" doc:"Ministerial score" format:"float" example:"65.7"`
+	MinisterialScore float64   `query:"ministerial_score" json:"ministerial_score" doc:"Ministerial score" format:"float64" example:"65.7"`
 	ScientistID      uuid.UUID `query:"scientist_id" json:"scientist_id" doc:"Scientist ID" format:"uuid" example:"8c4bfb01-3c0a-416c-a07c-a24ee52a8b2a"`
 }
 type DeleteBibliometric struct {
